+++ SipekSdk/SipekSdk.csproj
@@ -64,7 +64,7 @@
   <ItemGroup>
     <Compile Include="Common\AbstractFactory.cs" />
     <Compile Include="Common\AbstractWrapper.cs" />
-    <Compile Include="Common\CallControl\callManager.cs" />
+    <Compile Include="Common\CallControl\CallManager.cs" />
     <Compile Include="Common\CallControl\callStateMachine.cs" />
     <Compile Include="Common\CallControl\callStates.cs" />
     <Compile Include="Common\ICallLogInterface.cs" />
